Python_Lab_Project--main/ai/requirements.txt
fastapi
uvicorn
langchain
langchain-community 
langchain-core
langchain-chroma
langchain-google-genai
langchain-openai
chromadb
pydantic
python-dotenv
pandas
httpx
